split up concerns without altering functionality
